requirements.txt
absl-py==2.0.0
alabaster==0.7.13
anyio==4.0.0
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
atomicwrites==1.4.1
attrs==23.1.0
automium-web==0.1.1
automium==0.2.6
Babel==2.13.1
beautifulsoup4==4.12.2
black==23.11.0
bleach==6.1.0
blinker==1.7.0
boto3==1.28.80
botocore==1.31.80
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
curio==1.6
cycler==0.12.1
Cython==3.0.5
debugpy==1.8.0
decorator==5.1.1
deeplake==3.8.4
defusedxml==0.7.1
dill==0.3.7
Django==4.2.7
docrepr==0.2.0
docutils==0.18.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.13.1
flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
fsspec==2023.10.0
gast==0.4.0
google-auth-oauthlib==1.0.0
google-auth==2.23.3
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
hub==3.0.1
humbug==0.3.2
idna==3.4
image==1.5.33
ImageFilter==0.1.0
imageio==2.32.0
imagesize==1.4.1
imgaug==0.4.0
iniconfig==2.0.0
ipykernel==6.26.0
ipyparallel==8.6.1
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.20
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema-specifications==2023.7.1
jsonschema==4.19.2
jupyter-client==8.6.0
jupyter-core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter-server-terminals==0.4.4
jupyter-server==2.10.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.25.1
jupyterlab-widgets==3.0.9
jupyterlab==4.0.8
keras-tuner==1.4.6
keras==2.15.0
kiwisolver==1.4.5
kt-legacy==1.0.5
lazy-loader==0.3
legacy==0.1.7
libclang==16.0.6
lz4==4.3.2
Markdown==3.5
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
matplotlib==3.8.1
mistune==3.0.2
ml-dtypes==0.2.0
MouseInfo==0.1.3
mpmath==1.3.0
mrcnn==0.2
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook-shim==0.2.3
notebook==7.0.6
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
pathspec==0.11.2
Pillow==10.1.0
pip==21.3.1
platformdirs==3.11.0
pluggy==1.3.0
pox==0.3.3
ppft==1.7.6.7
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
pure-eval==0.2.2
py==1.11.0
pyasn1-modules==0.3.0
pyasn1==0.5.0
PyAutoGUI==0.9.54
pycocotools==2.0.7
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.16.1
PyJWT==2.8.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytest-asyncio==0.20.3
pytest==6.2.5
python-dateutil==2.8.2
python-json-logger==2.0.7
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
requests-oauthlib==1.3.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
rsa==4.9
s3transfer==0.7.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.2
setuptools==65.5.1
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
sphinx-rtd-theme==1.3.0
sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tensorboard-data-server==0.7.2
tensorboard==2.15.1
tensorflow-decision-forests==1.8.1
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.15.0
termcolor==2.3.0
terminado==0.17.1
testpath==0.6.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trio==0.23.1
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
utils==1.0.1
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
werkzeug==3.0.1
wget==3.2
wheel==0.38.4
widgetsnbextension==4.0.9
wrapt==1.14.1
wurlitzer==3.0.3
yolo-pyutils==0.1.36
